Added unit tests for measurement
